+++ CHANGES.rst
@@ -9,6 +9,7 @@ next
 - Log warning if not enough data is passed to fit() for a meaningful fit.
 - Better handle degenerate fits.
 - Allow missing observations in input multinomial data.
+- Avoid repeatedly rechecking validity of Gaussian covariance matrices.
 
 Version 0.2.2
 -------------